make the devices command cross platform
